+++ src/Gateway.php
@@ -347,6 +347,21 @@ class Gateway extends Core_Gateway {
 			$request->issuer = $payment->get_issuer();
 		}
 
+		// Billing email.
+		$billing_email = $payment->get_email();
+
+		/**
+		 * Filters the Mollie payment billing email used for bank transfer payment instructions.
+		 *
+		 * @since 2.2.0
+		 *
+		 * @param string|null $billing_email Billing email.
+		 * @param Payment     $payment       Payment.
+		 */
+		$billing_email = \apply_filters( 'pronamic_pay_mollie_payment_billing_email', $billing_email, $payment );
+
+		$request->set_billing_email( $billing_email );
+
 		// Due date.
 		try {
 			$due_date = new DateTime( sprintf( '+%s days', $this->config->due_date_days ) );