+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased][unreleased]
--
+- Added support for Mollie payment billing email.
 
 ## [2.1.2] - 2020-04-03
 - Fixed install issues on some specific WordPress installations.